requirements.txt
Django==3.2.4
psycopg2==2.9.1
djangorestframework==3.12.4
python-dateutil==2.8.1
django-multiselectfield==0.1.12
django-heroku==0.3.1
gunicorn==20.1.0
django-cors-headers==3.7.0
django-model-utils==4.1.1
django-simple-history==3.0.0
docxtpl==0.11.5
yadisk==1.2.14
pytz==2021.1
# test
parameterized==0.8.1